Some refactoring and updates to docs, especially splitting up sections about data files and added some examples. Also some more tests
