+++ BPt/__init__.py
@@ -13,6 +13,7 @@ from .main.input_operations import (Select, Duplicate, Pipe, Value_Subset,
                                     Intersection)
 from .main.BPtEvaluator import BPtEvaluator
 from . import p
+from pandas import from_pickle
 
 __author__ = "sahahn"
 __version__ = "2"
@@ -25,4 +26,4 @@ __all__ = ["Dataset", "Loader",
            "CV", "CVStrategy", "Intersection",
            "cross_validate", "cross_val_score", "evaluate",
            'p', "Feat_Selector", 'ModelPipeline', 'Pipeline',
-           'BPtEvaluator']
+           'BPtEvaluator', 'from_pickle']
