+++ doc/source/reference/api/BPt.BPtEvaluator.train_indices.rst
@@ -1,0 +1,6 @@
+BPt.BPtEvaluator.train\_indices
+===============================
+
+.. currentmodule:: BPt
+
+.. autoproperty:: BPtEvaluator.train_indices
