--- doc/source/reference/api/BPt.BPtEvaluator.train_indxs.rst
@@ -1,6 +1,0 @@
-BPt.BPtEvaluator.train\_indxs
-=============================
-
-.. currentmodule:: BPt
-
-.. autoproperty:: BPtEvaluator.train_indxs
