+++ doc/source/reference/api/BPt.BPtEvaluator.val_indices.rst
@@ -1,0 +1,6 @@
+BPt.BPtEvaluator.val\_indices
+=============================
+
+.. currentmodule:: BPt
+
+.. autoproperty:: BPtEvaluator.val_indices
