--- doc/source/reference/api/BPt.BPtEvaluator.val_indxs.rst
@@ -1,6 +1,0 @@
-BPt.BPtEvaluator.val\_indxs
-===========================
-
-.. currentmodule:: BPt
-
-.. autoproperty:: BPtEvaluator.val_indxs
