+++ doc/source/user_guide/saving_data.rst
@@ -5,3 +5,6 @@
 **************
 Saving Data
 **************
+
+The best way to save :class:`Dataset` is through the pandas base method :func:`pandas.DataFrame.to_pickle`.
+Likewise, you can load saved :class:`Dataset` with the pandas function :func:`pandas.load_pickle`.
