+++ docs/reference/api/BPt.evaluate.html
@@ -325,7 +325,7 @@ of numpy arrays with the raw predictions.</p>
 also will stored the corresponding ground
 truth labels under dictionary key ‘y_true’ in
 preds. Or to see corresponding subjects / index,
-check ‘val_indxs’ in the evaluator.</p>
+check ‘val_indices’ in the evaluator.</p>
 </dd>
 <dt><strong>store_estimators</strong><span class="classifier">bool, optional</span></dt><dd><p>Skip.</p>
 <div class="highlight-default notranslate"><div class="highlight"><pre><span></span><span class="n">default</span> <span class="o">=</span> <span class="kc">True</span>
